cast offset stuff as int16 -.-'

+++ Assignment2/mymalloc.c
@@ -272,9 +272,9 @@ void * myallocate(size_t size, char * b, int a, int id){
 		
 		if (whichMemory == memory + OS_SIZE){
 
-			headerPage = (headerAddress - (memory + OS_SIZE))/pageSize;
+			headerPage = (int16_t) (headerAddress - (memory + OS_SIZE))/pageSize;
 				
-			footerPage = (footerAddress - (memory + OS_SIZE))/pageSize;
+			footerPage = (int16_t) (footerAddress - (memory + OS_SIZE))/pageSize;
 			
 			void * pagePointer;
 			int16_t i = 0;
